refactor: 특정 제품의 사용자 연차와 직군의 총개수를 반환하는 Repository 패키지 변경
