+++ backend/src/main/java/com/woowacourse/f12/domain/product/ProductRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Long>, ProductRepositoryCustom {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
     Slice<Product> findPageBy(Pageable pageable);
