+++ backend/src/main/java/com/woowacourse/f12/domain/review/ReviewRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface ReviewRepository extends JpaRepository<Review, Long> {
+public interface ReviewRepository extends JpaRepository<Review, Long>, ReviewRepositoryCustom {
 
     @Query("select r from Review r where r.product.id = :productId")
     Slice<Review> findPageByProductId(Long productId, Pageable pageable);
