+++ backend/src/main/java/com/woowacourse/f12/domain/review/ReviewRepositoryCustomImpl.java
@@ -1,20 +1,21 @@
-package com.woowacourse.f12.domain.product;
+package com.woowacourse.f12.domain.review;
 
 import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import com.woowacourse.f12.domain.member.QMember;
-import com.woowacourse.f12.domain.review.QReview;
+import com.woowacourse.f12.domain.product.Product;
+import com.woowacourse.f12.domain.product.QProduct;
 import java.util.List;
 import org.springframework.data.jpa.repository.support.QuerydslRepositorySupport;
 
-public class ProductRepositoryCustomImpl extends QuerydslRepositorySupport implements ProductRepositoryCustom {
+public class ReviewRepositoryCustomImpl extends QuerydslRepositorySupport implements ReviewRepositoryCustom {
 
     private final QProduct product = QProduct.product;
     private final QReview review = QReview.review;
     private final QMember member = QMember.member;
     private final JPAQueryFactory jpaQueryFactory;
 
-    public ProductRepositoryCustomImpl(final JPAQueryFactory jpaQueryFactory) {
+    public ReviewRepositoryCustomImpl(final JPAQueryFactory jpaQueryFactory) {
         super(Product.class);
         this.jpaQueryFactory = jpaQueryFactory;
     }
