+++ backend/src/test/java/com/woowacourse/f12/domain/review/ReviewRepositoryTest.java
@@ -1,8 +1,15 @@
 package com.woowacourse.f12.domain.review;
 
+import static com.woowacourse.f12.domain.member.CareerLevel.JUNIOR;
+import static com.woowacourse.f12.domain.member.CareerLevel.SENIOR;
+import static com.woowacourse.f12.domain.member.JobType.BACKEND;
+import static com.woowacourse.f12.domain.member.JobType.MOBILE;
+import static com.woowacourse.f12.support.MemberFixtures.CORINNE;
+import static com.woowacourse.f12.support.MemberFixtures.MINCHO;
 import static com.woowacourse.f12.support.ProductFixture.KEYBOARD_1;
 import static com.woowacourse.f12.support.ProductFixture.KEYBOARD_2;
-import static com.woowacourse.f12.support.MemberFixtures.CORINNE;
+import static com.woowacourse.f12.support.ReviewFixtures.REVIEW_RATING_1;
+import static com.woowacourse.f12.support.ReviewFixtures.REVIEW_RATING_2;
 import static com.woowacourse.f12.support.ReviewFixtures.REVIEW_RATING_4;
 import static com.woowacourse.f12.support.ReviewFixtures.REVIEW_RATING_5;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -14,6 +21,7 @@ import com.woowacourse.f12.domain.member.Member;
 import com.woowacourse.f12.domain.member.MemberRepository;
 import com.woowacourse.f12.domain.product.Product;
 import com.woowacourse.f12.domain.product.ProductRepository;
+import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import org.junit.jupiter.api.Test;
@@ -143,6 +151,71 @@ class ReviewRepositoryTest {
 
         // then
         assertThat(actual).isTrue();
+    }
+
+    @Test
+    void 제품에_대한_사용자_연차의_총_개수를_반환한다() {
+        // given
+        Product product = 제품_저장(KEYBOARD_1.생성());
+
+        Member corinne = CORINNE.생성();
+        corinne.updateCareerLevel(SENIOR);
+        corinne.updateJobType(MOBILE);
+        corinne = memberRepository.save(corinne);
+
+        Member mincho = MINCHO.생성();
+        mincho.updateCareerLevel(JUNIOR);
+        mincho.updateJobType(BACKEND);
+        mincho = memberRepository.save(mincho);
+
+        리뷰_저장(REVIEW_RATING_2.작성(product, corinne));
+        리뷰_저장(REVIEW_RATING_1.작성(product, mincho));
+
+        // when
+        List<CareerLevelCount> careerLevelCounts = reviewRepository.findCareerLevelCountByProductId(
+                product.getId());
+
+        // then
+        assertThat(careerLevelCounts).usingRecursiveFieldByFieldElementComparator()
+                .hasSize(2)
+                .containsOnly(
+                        new CareerLevelCount(JUNIOR, 1),
+                        new CareerLevelCount(SENIOR, 1)
+                );
+    }
+
+    @Test
+    void 제품에_대한_사용자_직군의_총_개수를_반환한다() {
+        // given
+        Product product = 제품_저장(KEYBOARD_1.생성());
+
+        Member corinne = CORINNE.생성();
+        corinne.updateCareerLevel(SENIOR);
+        corinne.updateJobType(MOBILE);
+        corinne = memberRepository.save(corinne);
+
+        Member mincho = MINCHO.생성();
+        mincho.updateCareerLevel(JUNIOR);
+        mincho.updateJobType(BACKEND);
+        mincho = memberRepository.save(mincho);
+
+        리뷰_저장(REVIEW_RATING_2.작성(product, corinne));
+        리뷰_저장(REVIEW_RATING_1.작성(product, mincho));
+
+        // when
+        List<JobTypeCount> jobTypeCounts = reviewRepository.findJobTypeCountByProductId(product.getId());
+
+        // then
+        assertThat(jobTypeCounts).usingRecursiveFieldByFieldElementComparator()
+                .hasSize(2)
+                .containsOnly(
+                        new JobTypeCount(MOBILE, 1),
+                        new JobTypeCount(BACKEND, 1)
+                );
+    }
+
+    private Product 제품_저장(Product product) {
+        return productRepository.save(product);
     }
 
     private Review 리뷰_저장(Review review) {
